+++ dbbackup/storage.py
@@ -5,7 +5,6 @@ Utils for handle files.
 import logging
 
 from django.core.exceptions import ImproperlyConfigured
-from django.core.files.storage import DefaultStorage
 
 from . import settings, utils
 
@@ -287,13 +286,13 @@ class Storage:
             self.delete_file(filename)
 
 
-def get_storage_class(path: str | None = None):
+def get_storage_class(path=None):
     """
     Return the configured storage class.
     
     :param path: Path in Python dot style to module containing the storage
-                    class. If empty settings.DBBACKUP_STORAGE will be used.
-    :type path: str
+                    class. If empty, the default storage class will be used.
+    :type path: str or None
         
     :returns: Storage class
     :rtype: :class:`django.core.files.storage.Storage`
@@ -304,10 +303,9 @@ def get_storage_class(path: str | None = None):
         # this is a workaround to keep compatibility with Django >= 5.1 (django.core.files.storage.get_storage_class is removed)
         return import_string(path)
     
-    # TODO: Use django.core.files.storage.storages dict on Django >= 5.1
-    # This would require a change in the dbbackup settings to make use of the the new STORAGES dict settings
-    # like so: STORAGES = { "dbbackup": { "BACKEND": "django.core.files.storage.FileSystemStorage", "OPTIONS": {...} } }
-    # and then returning storages['dbbackup'] here
-    
-    return DefaultStorage
-    
+    try:
+        from django.core.files.storage import DefaultStorage
+        return DefaultStorage
+    except Exception:
+        from django.core.files.storage import get_storage_class
+        return get_storage_class()