+++ dbbackup/tests/settings.py
@@ -63,6 +63,24 @@ DBBACKUP_STORAGE_OPTIONS = dict(
     ]
 )
 
+# For testing the new storages setting introduced in Django 4.2
+# STORAGES = {
+#     "default": {
+#         "BACKEND": "django.core.files.storage.FileSystemStorage",
+#         "OPTIONS": {},
+#     },
+#     "dbbackup": {
+#         "BACKEND": os.environ.get("STORAGE", "dbbackup.tests.utils.FakeStorage"),
+#         "OPTIONS": dict(
+#             [
+#                 keyvalue.split("=")
+#                 for keyvalue in os.environ.get("STORAGE_OPTIONS", "").split(",")
+#                 if keyvalue
+#             ]
+#         )
+#     }
+# }
+
 LOGGING = {
     "version": 1,
     "disable_existing_loggers": False,